feat: 모바일 화면 touchMove 드래그 기능 모듈화
