+++ src/util/index.ts
@@ -1,3 +1,4 @@
 export * from './date';
 export * from './pagination';
 export * from './getDateTypeToString';
+export * from './touchMoveDrag';
